Amend HL ADoc Extension v1.2.0
Amend the text mentioning preferences search priorities being different
in Windows and Linux/Mac, both in the documents and the extension source
comments. Version number of the extension unchanged, since this only
affects comments, not functionality (and v1.2.0 hasn't yet been included
in a release).

For more details, see discussion in merge request !105:

https://gitlab.com/saalen/highlight/merge_requests/105